requirements.txt
qrcode
aiohttp
ping3
requests
discordpy
